Check in the scientific name string

Identifies open nomenclature (aff., cf.) in species scientific name, classification under species level (var. and subsp.), and common mistakes in the a species scientific name. It creates a new column with the original string and the new suggested name. It also flags problematic names (character string with numbers, authors, wrong case, or other names besides genus and epithet etc).

check_string(species = NULL)

Arguments

species
species name in a raw format

Value

Data frame with `verbatimSpecies` as in the original input, `speciesStatus` with the flags in original data and `species` with a suggestion for a more correct species name. See Details for a description of flags in the column `speciesStatus`.

Details

Possible flags returned in `species_status`:

possibly_ok

scientific name following the expected pattern Genus epithet

not_Genus_epithet_format

scientific name not following the expected pattern Genus epithet

variety

species scientific name with variety

subspecies

species scientific name with subspecies

form

species scientific name with form

conferre

open nomenclature cf. in species scientific name

affinis

open nomenclature aff. in species scientific name

name_w_authors

species scientific name has authors

not_name_has_digits

species scientific name has digits, not a valid name

indet

species identified only at genus level

family_as_genus

family as genus, not a valid name

order_as_genus

order as genus, not a valid name

species_nova

species name contains an indication of a new species, possibly not yet a valid name

non_ascii

species name has non ASCII characters, not a valid name

#'
hybrid_species

hybrid species

Author

Sara Mortara & Andrea Sánchez-Tapia

Examples

check_string("Lindsaea lancea var. falcata")

#>                verbatimSpecies speciesStatus         species
#> 1 Lindsaea lancea var. falcata       variety Lindsaea lancea

check_string(c("Lindsaea lancea var. falcata", "Asplenium Aff. truncorum"))

#>                verbatimSpecies speciesStatus             species
#> 1 Lindsaea lancea var. falcata       variety     Lindsaea lancea
#> 2     Asplenium Aff. truncorum       affinis Asplenium truncorum

check_string("Asplenium sp.")

#>   verbatimSpecies speciesStatus       species
#> 1   Asplenium sp.         indet Asplenium sp.